+++ NEWS.md
@@ -28,6 +28,7 @@
 * Updated `scatterplot()` function to allow plotting of multiple parameters
 * Updated `match_worms_taxa()` function to clean taxon names from problematic special characters before being passed to API call
 * Added `plot_leaflet` argument to functions `check_station_distance()` and `check_onland()`
+* Added `utv` argument to functions `get_shark_options()`, `get_shark_data()`, `get_shark_datasets()` and `get_shark_table_counts()`.
 * Added `add_rank_to_hierarchy` argument to the `add_worms_taxonomy()` function
 * `match_worms_taxa()` now handles bulk API requests using the `bulk` argument
 * Add SHARK4R Bio-QC Tool Shiny App to bundle, with improved performance (initialized by `run_qc_app()`)